requirements.txt
watchdog==6.0.0
python-dotenv==1.2.1
